+++ app/src/main/java/com/example/mywork/domain/work/Work.kt
@@ -9,5 +9,5 @@ data class Work(
     val organization: Organization,
     val worker: Worker,
     val description: String,
-    val time: Long
+    val time: Double
 )
